refactor: Refatoração página DetalhesImagens
