+++ src/components/FullScreenModal/FullScreenModal.styles.ts
@@ -1,0 +1,36 @@
+import { StyleSheet, Dimensions } from 'react-native';
+
+const { width, height } = Dimensions.get('window');
+
+export const styles = StyleSheet.create({
+    fullScreenContainer: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: 'rgba(0, 0, 0, 0.9)',
+    },
+    fullScreenImageWrapper: {
+        width: width * 0.9, 
+        justifyContent: 'center',
+        alignItems: 'center',
+        marginHorizontal: width * 0.05,
+    },
+    fullScreenImage: {
+        width: '100%',
+        height: height * 0.9,
+        borderRadius: 10,
+    },
+    closeButton: {
+        position: 'absolute',
+        top: 40,
+        right: 20,
+        backgroundColor: 'rgba(255, 255, 255, 0.8)',
+        borderRadius: 15,
+        padding: 10,
+    },
+    closeButtonText: {
+        fontSize: 18,
+        fontWeight: 'bold',
+        color: '#333',
+    },
+});
