+++ src/components/ImageDetailsCard/ImageDetailsCard.styles.ts
@@ -1,0 +1,28 @@
+import { StyleSheet, Dimensions } from 'react-native';
+const { width } = Dimensions.get('window');
+
+export const styles = StyleSheet.create({
+    card: {
+        width: width * 0.9,
+        padding: 20,
+        borderRadius: 10,
+        backgroundColor: '#f9f9f9',
+        elevation: 5,
+        shadowColor: '#000',
+        shadowOffset: { width: 0, height: 2 },
+        shadowOpacity: 0.3,
+        shadowRadius: 3,
+        alignSelf: 'center',
+        marginVertical: 20,
+    },
+    title: {
+        fontSize: 22,
+        fontWeight: 'bold',
+        marginBottom: 10,
+        textAlign: 'center',
+    },
+    text: {
+        fontSize: 16,
+        marginBottom: 5,
+    },
+});
