+++ src/pages/DetalhesImagens/DetalhesImagens.styles.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: 'white',
+    },
+});
